backend/requirements.txt
fastapi~=0.115.0
motor==3.6.0
pydantic==2.9.2
asyncpg~=0.30.0
SQLAlchemy==2.0.36
GeoAlchemy2~=0.15.2
aiohttp~=3.10.10
python-dotenv~=1.0.1
PyJWT~=2.3.0
bcrypt
mypy
pytest
aiobotocore
botocore
pytest-asyncio
sqlalchemy[asyncio]